Move Feed Image Data Loader Decorator to production
